feature: 1.基本实现了视频详细界面的数据展示 视频播放 点击事件 2.社区->推荐:设置了视频点击事件 即跳转到1中实现的界面 fix: 1.修改了第三方库:HTextView的版本

+++ app/src/main/java/com/greenhand/cooperativework/bean/VideoReplyBean.kt
@@ -31,7 +31,7 @@ data class VideoReplyBean(
         val liked: Boolean,
         val message: String,
         val parentReply: Any,
-        val parentReplyId: Int,
+        val parentReplyId: Long,
         val replyStatus: String,
         val rootReplyId: Long,
         val sequence: Int,
